requirements.txt
pandas
psycopg2-binary
scikit-learn
python-dotenv
schedule
gunicorn
